feat: Add PancakeSwap V3 contracts and update project configuration
- Installed PancakeSwap V3 contracts as a git submodule
- Updated foundry.toml with new remapping for PancakeSwap V3
- Removed example Counter contract and test files
- Minor formatting and whitespace improvements in existing files

+++ src/interfaces/IDEX.sol
@@ -19,29 +19,30 @@ interface IDEXPair {
 // DEX Router 接口
 interface IDEXRouter {
     function factory() external pure returns (address);
-    
-    function getAmountsOut(uint amountIn, address[] calldata path) external view returns (uint[] memory amounts);
-    
+
+    function getAmountsOut(uint256 amountIn, address[] calldata path)
+        external
+        view
+        returns (uint256[] memory amounts);
+
     function swapExactTokensForTokens(
-        uint amountIn,
-        uint amountOutMin,
+        uint256 amountIn,
+        uint256 amountOutMin,
         address[] calldata path,
         address to,
-        uint deadline
-    ) external returns (uint[] memory amounts);
-    
-    function swapExactETHForTokens(
-        uint amountOutMin,
-        address[] calldata path,
-        address to,
-        uint deadline
-    ) external payable returns (uint[] memory amounts);
-    
+        uint256 deadline
+    ) external returns (uint256[] memory amounts);
+
+    function swapExactETHForTokens(uint256 amountOutMin, address[] calldata path, address to, uint256 deadline)
+        external
+        payable
+        returns (uint256[] memory amounts);
+
     function swapExactTokensForETH(
-        uint amountIn,
-        uint amountOutMin,
+        uint256 amountIn,
+        uint256 amountOutMin,
         address[] calldata path,
         address to,
-        uint deadline
-    ) external returns (uint[] memory amounts);
-} 
+        uint256 deadline
+    ) external returns (uint256[] memory amounts);
+}
